# docs/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: API de Finanzas Inteligentes
  version: 1.0.0-draft
  description: Contrato ejecutable derivado de CONVENCIONES.md, ENDPOINTS.md, CONTRATOS.md y TRAZABILIDAD_RF.md.
servers:
- url: https://api.finanzas.example/api/v1
  description: Producción (referencial)
- url: http://localhost:8080/api/v1
  description: Desarrollo local
security:
- bearerAuth: []
...
paths:
...
  "/proyecciones-gastos":
    get:
      tags:
      - Predicciones, alertas y score
      summary: GET /proyecciones-gastos
      operationId: getProyeccionesGastos
      x-requisitos-funcionales:
      - RF-16
      x-etapa: 6
      x-pruebas-contrato:
      - CT-RF-16
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: ambito
        in: query
        required: false
        schema:
          type: string
      - name: periodo
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProyeccionResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '403':
          "$ref": "#/components/responses/Problem403"
        '422':
          "$ref": "#/components/responses/Problem422"
        '500':
          "$ref": "#/components/responses/Problem500"
...
  "/alertas-financieras":
    get:
      tags:
      - Predicciones, alertas y score
      summary: GET /alertas-financieras
      operationId: getAlertasFinancieras
      x-requisitos-funcionales:
      - RF-17
      x-etapa: 6
      x-pruebas-contrato:
      - CT-RF-17
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - "$ref": "#/components/parameters/Cursor"
      - "$ref": "#/components/parameters/Limite"
      - name: nivel
        in: query
        required: false
        schema:
          type: string
      - name: leida
        in: query
        required: false
        schema:
          type: boolean
      - name: desde
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: hasta
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PaginaAlertaResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '500':
          "$ref": "#/components/responses/Problem500"
  "/alertas-financieras/{alertaId}":
    get:
      tags:
      - Predicciones, alertas y score
      summary: GET /alertas-financieras/{alertaId}
      operationId: getAlertasFinancierasByAlertaId
      x-requisitos-funcionales:
      - RF-17
      x-etapa: 6
      x-pruebas-contrato:
      - CT-RF-17
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: alertaId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AlertaResponse"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '500':
          "$ref": "#/components/responses/Problem500"
    patch:
      tags:
      - Predicciones, alertas y score
      summary: PATCH /alertas-financieras/{alertaId}
      operationId: patchAlertasFinancierasByAlertaId
      x-requisitos-funcionales:
      - RF-17
      x-etapa: 6
      x-pruebas-contrato:
      - CT-RF-17
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      - name: alertaId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IfMatch"
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
            ETag:
              "$ref": "#/components/headers/ETag"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AlertaResponse"
        '400':
          "$ref": "#/components/responses/Problem400"
        '401':
          "$ref": "#/components/responses/Problem401"
        '404':
          "$ref": "#/components/responses/Problem404"
        '412':
          "$ref": "#/components/responses/Problem412"
        '500':
          "$ref": "#/components/responses/Problem500"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PatchAlertasFinancierasByAlertaIdRequest"
  "/score-financiero":
    get:
      tags:
      - Predicciones, alertas y score
      summary: GET /score-financiero
      operationId: getScoreFinanciero
      x-requisitos-funcionales: []
      x-etapa: 6
      x-pruebas-contrato:
      - CT-SCORE
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/CorrelationId"
      responses:
        '200':
          description: Operación exitosa.
          headers:
            X-Correlation-Id:
              "$ref": "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ScoreResponse"
        '401':
          "$ref": "#/components/responses/Problem401"
        '422':
          "$ref": "#/components/responses/Problem422"
        '500':
          "$ref": "#/components/responses/Problem500"
      x-capacidad: CT-SCORE
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 100
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: ^\".+\"$
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        maxLength: 500
    Limite:
      name: limite
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  headers:
    CorrelationId:
      description: Identificador de correlación de la solicitud.
      schema:
        type: string
    ETag:
      description: Versión opaca del recurso.
      schema:
        type: string
...
  schemas:
...
    ProyeccionResponse:
      type: object
      additionalProperties: false
      properties:
        ambito:
          type: string
          example: privado
        mesesHistorial:
          type: integer
          format: int64
          example: 2
        preliminar:
          type: boolean
          example: true
        gastoProyectado:
          type: integer
          format: int64
          example: 3850000
        balanceProyectado:
          type: integer
          format: int64
          example: 4050000
        categoriaMayorCrecimiento:
          type: string
          example: Transporte
        nivelRiesgo:
          type: string
          example: bajo
        versionModelo:
          type: string
          example: gastos-v1
        periodo:
          type: object
          additionalProperties: false
          properties:
            desde:
              type: string
              format: date
              example: '2026-07-01'
            hasta:
              type: string
              format: date
              example: '2026-12-31'
          required:
          - desde
          - hasta
        categorias:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              nombre:
                type: string
                example: Transporte
              montoProyectado:
                type: integer
                format: int64
                example: 620000
              variacion:
                type: number
                format: double
                example: 0.28
            required:
            - nombre
            - montoProyectado
            - variacion
        historial:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              periodo:
                type: string
                example: 2026-06
              proyectado:
                type: integer
                format: int64
                example: 3600000
              real:
                type: integer
                format: int64
                example: 3550000
            required:
            - periodo
            - proyectado
            - real
        generadoEn:
          type: string
          format: date-time
          example: '2026-07-22T18:30:00Z'
      required:
      - ambito
      - mesesHistorial
      - preliminar
      - gastoProyectado
      - balanceProyectado
      - categoriaMayorCrecimiento
      - nivelRiesgo
      - versionModelo
      - periodo
      - categorias
      - historial
      - generadoEn
    AlertaResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        titulo:
          type: string
          example: Gasto en transporte por encima del promedio
        mensaje:
          type: string
          example: Transporte representa 18% de tus gastos registrados.
        nivel:
          type: string
          example: advertencia
        fecha:
          type: string
          format: date-time
          example: '2026-07-22T18:30:00Z'
        queOcurrio:
          type: string
          example: El gasto supera el promedio reciente.
        datosUtilizados:
          type: string
          example: Movimientos de los últimos 90 días.
        impacto:
          type: string
          example: El presupuesto podría agotarse antes de fin de mes.
        recomendacion:
          type: string
          example: Revisá los traslados recientes.
        leida:
          type: boolean
          example: false
        version:
          type: integer
          format: int64
          example: 1
      required:
      - id
      - titulo
      - mensaje
      - nivel
      - fecha
      - queOcurrio
      - datosUtilizados
      - impacto
      - recomendacion
      - leida
      - version
    ScoreResponse:
      type: object
      additionalProperties: false
      properties:
        score:
          type: integer
          format: int64
          example: 78
        estado:
          type: string
          example: equilibrado
        versionAlgoritmo:
          type: string
          example: score-v1
        periodo:
          type: object
          additionalProperties: false
          properties:
            desde:
              type: string
              format: date
              example: '2026-04-01'
            hasta:
              type: string
              format: date
              example: '2026-06-30'
          required:
          - desde
          - hasta
        factoresPositivos:
          type: array
          items:
            type: string
            example: Balance mensual positivo
        factoresNegativos:
          type: array
          items:
            type: string
            example: Aumento en gastos variables
        historial:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              periodo:
                type: string
                example: 2026-06
              score:
                type: integer
                format: int64
                example: 76
            required:
            - periodo
            - score
        recomendaciones:
          type: array
          items:
            type: string
            example: Mantené un fondo de emergencia
      required:
      - score
      - estado
      - versionAlgoritmo
      - periodo
      - factoresPositivos
      - factoresNegativos
      - historial
      - recomendaciones
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - codigo
      - correlationId
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type:
          - string
          - 'null'
        instance:
          type:
          - string
          - 'null'
        codigo:
          type: string
          pattern: "^[a-z0-9_]+$"
        correlationId:
          type: string
        errores:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
    Paginacion:
      type: object
      additionalProperties: false
      required:
      - siguienteCursor
      - hayMas
      - limite
      properties:
        siguienteCursor:
          type:
          - string
          - 'null'
        hayMas:
          type: boolean
        limite:
          type: integer
          minimum: 1
          maximum: 100
...
    PaginaAlertaResponse:
      type: object
      additionalProperties: false
      required:
      - datos
      - paginacion
      properties:
        datos:
          type: array
          items:
            "$ref": "#/components/schemas/AlertaResponse"
        paginacion:
          "$ref": "#/components/schemas/Paginacion"
...
    PatchAlertasFinancierasByAlertaIdRequest:
      type: object
      additionalProperties: false
      properties:
        leida:
          type: boolean
        archivada:
          type: boolean
      minProperties: 1
...
  responses:
    Problem400:
      description: Solicitud o parámetros inválidos.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem401:
      description: Autenticación ausente o inválida.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem403:
      description: Permiso insuficiente.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
    Problem404:
      description: Recurso inexistente o no visible.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem412:
      description: ETag desactualizado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem422:
      description: Regla de negocio no satisfecha.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...
    Problem500:
      description: Error interno no controlado.
      headers:
        X-Correlation-Id:
          "$ref": "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/ProblemDetails"
...